map_kind: apply inverse mapping to *rule.File provided to GenerateRules
A language module must process existing rules, and for it to remain ignorant of
Mapped Kinds (a design goal for that feature), it must read the input BUILD file
as if the mapping hadn't been applied.

Unfortunately, this issue is not possible to trigger with the Go rules, so a fix
plus test may be difficult to merge upstream.